Add support for local whispercpp
Now localwhisper will expect a whispercpp server instead of the old faster_whisper Python server.

Overall whispercpp is a better option, it will work well both for CPU and GPU (there are separate whispercpp binaries for CPU and GPU)
It also supports optimized models like distil-whisper and such.
The caveat is that whispercpp only accepts 16000Hz input so we need to run ffmpeg to convert the .wav before sending it via HTTP.
That should add only minimal latency like 50-200ms. It's worth it because distil-whisper models are insanely fast.