Add xtts2 TTS support
To be used with https://github.com/daswer123/xtts-api-server

The xtts-api-server must be run from /home/dwemer and a 'input_samples' folder must be there with the file scarlett24000.wav in there.
This is the input sample that xtts2 will try to clone and use for TTS. In my case I had a few seconds of female voice speaking.

The server should be run like this:

python -m xtts_api_server -sf input_samples

Then change in conf.php:

$TTSFUNCTION='xtts2';

Then it should work with Herika. Make sure you have NVIDIA GPU. The TTS is slow because I think the server doesnot use deepspeed for CUDA.
I will try to get it working with deepspeed. In my case it took like 10seconds to generate a response on a 4090.

For the record here are the install steps that I had to do in DwemerDistro shell to get the server to work:

  367  . venv/bin/activate
  368  pip install xtts-api-server
  369  pip install torch==2.1.1+cu118 torchaudio==2.1.1+cu118 --index-url https://download.pytorch.org/whl/cu118
  370  pip install xtts-api-server
  371  sudo apt-get install portaudio19-dev
  372  pip install xtts-api-server